Write code for operate() and create = addEventListener
Wrote an addEveentListener for when the = button gets clicked. This will trigger the operate
function, where first the number strings are converted to numbers, and then the correct
function is triggered based on what the operator equals (+, -, *, or /). Currently trying to
fix an issue of the operate function ONLY performing addition, even when another operate
button is clicked