Style calculator and fix logix error
Have completed a good amount ofthe stylingin the CSS file, mostley focusing on expanding
button size and adjusting the panel of buttons to fit calculator. ALso noticed an
error in the calculator logic when any operator button is pressed before any numbers.
Fixed so that this has no effect if done before the number buttons